refactor: score math answers in a single module
The value comparison landed in its own module for tidiness, which left a
588-line file with exactly one caller and one public function nothing but its
own test used. Both concerns are one thing -- scoring a math answer -- so they
live together, and the test module follows the module.
